macos compile warnings

+++ arx/src/dbf_from.c
@@ -2,6 +2,7 @@
  * write dbf-structure file to stdout
  */
 #include <stdio.h>
+#include <string.h>
 #include "arx_def.h"
 #include "dbf_def.h"
 
@@ -13,7 +14,7 @@ struct FIELD_RECORD *f;
 if (m == 1) {
 printf ("adbfs := { ;\n");
 for (fld=1, f = d->fields_ptr; fld<=d->fields; fld++, f++)
-	printf ("  { '%s'%*s, '%c', %2d, %1d }%c ;\n", f->name, 11-strlen(f->name), "", f->typ, f->len, f->dec, fld==d->fields ? ' ' : ',');
+	printf ("  { '%s'%*s, '%c', %2d, %1d }%c ;\n", f->name, 11-(int)strlen(f->name), "", f->typ, f->len, f->dec, fld==d->fields ? ' ' : ',');
 printf ("  }\n");
 } else {
 printf ("FIELD_NAME\tFIELD_TYPE\tFIELD_LEN\tFIELD_DEC\n");
